car-rec/requirements.txt
-i https://pypi.org/simple
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.11.8
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cmake==3.18.2.post1
distlib==0.3.1
dlib==19.21.0
filelock==3.0.12
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.2; python_version >= '3.6'
google-auth==1.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
kafka-python==2.0.2
keras-preprocessing==1.1.2
markdown==3.3.3; python_version >= '3.6'
numpy==1.18.5; python_version >= '3.5'
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opt-einsum==3.3.0; python_version >= '3.5'
pillow==8.0.1; python_version >= '3.6'
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.24.0
rsa==4.6; python_version >= '3.5'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-plugin-wit==1.7.0
tensorboard==2.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-estimator==2.3.0
tensorflow==2.3.1
termcolor==1.1.0
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.35.1; python_version >= '3'
wrapt==1.12.1
